+++ tool/tuffixlib.py
@@ -1,17 +1,13 @@
+
 ################################################################################
 # imports
 ################################################################################
 
 # standard library
-import io, json, os, pathlib, sys, unittest, requests
+import io, json, os, pathlib, sys, unittest
 
 # packages
-import apt.cache, packaging.version, apt.debfile
-
-# our lib imports
-
-from TuffixExceptions import *
-from TuffixLang.Fetch import Fetch 
+import apt.cache, packaging.version
 
 ################################################################################
 # constants
@@ -23,6 +19,37 @@ STATE_PATH = pathlib.Path('/var/lib/tuffix/state.json')
 
 KEYWORD_MAX_LENGTH = 8
 
+################################################################################
+# exception types
+################################################################################
+
+# base types for exceptions that include a string message
+class MessageException(Exception):
+    def __init__(self, message):
+        if not (isinstance(message, str)):
+            raise ValueError
+        self.message = message
+
+# commandline usage error
+class UsageError(MessageException):
+    def __init__(self, message):
+        super().__init__(message)
+
+# problem with the environment (wrong OS, essential shell command missing, etc.)
+class EnvironmentError(MessageException):
+    def __init__(self, message):
+        super().__init__(message)
+
+# issue reported by the `status` command, that's at the level of a fatal error
+class StatusError(MessageException):
+    def __init__(self, message):
+        super().__init__(message)
+
+# issue reported by the `status` command, that's at the level of a nonfatal
+# warning
+class StatusWarning(MessageException):
+    def __init__(self, message):
+        super().__init__(message)
 
 ################################################################################
 # configuration
@@ -269,7 +296,91 @@ def all_commands(build_config):
 # keywords
 ################################################################################
 
+class AbstractKeyword:
+    def __init__(self, build_config, name, description):
+        if not (isinstance(build_config, BuildConfig) and
+                isinstance(name, str) and
+                len(name) <= KEYWORD_MAX_LENGTH and
+                isinstance(description, str)):
+            raise ValueError
+        self.name = name
+        self.description = description
 
+    def add(self):
+        raise NotImplementedError
+        
+    def remove(self):
+        raise NotImplementedError
+
+# Keyword names may begin with a course code (digits), but Python
+# identifiers may not. If a keyword name starts with a digit, prepend
+# the class name with C (for Course).
+
+class BaseKeyword(AbstractKeyword):
+
+    # TODO: install Atom, googletest, g++?
+    
+    packages = ['build-essential',
+                'clang',
+                'clang-tidy',
+                'clang-format']
+    
+    def __init__(self, build_config):
+        super().__init__(build_config,
+                         'base',
+                         'CPSC 120-121-131-301 C++ development environment')
+        
+    def add(self):
+        add_deb_packages(self.packages)
+        
+    def remove(self):
+        remove_deb_packages(self.packages)
+
+class C439Keyword(AbstractKeyword):
+
+    packages = ['minisat2']
+    
+    def __init__(self, build_config):
+        super().__init__(build_config, '439', 'CPSC 439')
+         
+    def add(self):
+        add_deb_packages(self.packages)
+        
+    def remove(self):
+        remove_deb_packages(self.packages)
+
+class LatexKeyword(AbstractKeyword):
+    packages = ['texlive-full']
+
+    def __init__(self, build_config):
+        super().__init__(build_config,
+                         'latex',
+                         'LaTeX typesetting environment (large)')
+         
+    def add(self):
+        add_deb_packages(self.packages)
+        
+    def remove(self):
+        remove_deb_packages(self.packages)
+
+# TODO: more keywords...
+
+def all_keywords(build_config):
+    if not isinstance(build_config, BuildConfig):
+        raise ValueError
+    # alphabetical order, but put digits after letters
+    return [ BaseKeyword(build_config),
+             LatexKeyword(build_config),
+             C439Keyword(build_config) ]
+
+def find_keyword(build_config, name):
+    if not (isinstance(build_config, BuildConfig) and
+            isinstance(name, str)):
+        raise ValueError
+    for keyword in all_keywords(build_config):
+        if keyword.name == name:
+            return keyword
+    raise UsageError('unknown keyword "' + name + '", see valid keyword names with $ tuffix list')
 
 ################################################################################
 # system probing functions (gathering info about the environment)
